dispatch handler based on connection version
Instead of having a single fixed table of handlers,
have one per protocol version.  That is, instead of
just table[i](args), we call table[vers][i](args).

The initial version is already always zero, so bump
the version index numbers up a bit so that the first
valid version is 1.  Set up the 0'th table to accept
only a Tversion operation.  This way we require a
Tversion to complete before we will allow any additional
operations, as described in the protocol documentation.