handle L9P_ORDWR
Inspired by ailispaw's changes (but simplified by cheating).

Add some comments as to what is going on.

Also, mark getpwuid/getgrouplist calls as not-thread-safe,
for now.  It's not easy to fix getgrouplist since it's a netdb
dispatch in libc, in FreeBSD, and there is no reentrant POSIX
variant available.  Will have to study this to see if just using
my r_getpwent() variant suffices, or if we need a thread lock.