example/server.c: report any failure
If l9p_start_server() returns a nonzero value, use err() to
exit.  (Perhaps should just use errx(); it's not clear errno
will necessarily be valid.)

Note that l9p_start_server() currently never returns 0, so
this is still pretty crude, but at least we catch the case of
running as non-root and hence not able to bind the socket.