utils.c: describe all the new .L ops
This includes a function to pack and unpack variable-length
directory entries, though it will not be used until we wire
up readdir.

Decode all the new ops for debugging.  Variable length direntry
decode is ifdef-ed out since it needs internal changes and is
of somewhat dubious value.

Need to decode and encode for wire format will go in as the
ops are hooked up.