fix readdir on large directories
The backend readdir() code assumed that seekdir() took negative
values to seek backwards, which is not the case.

While here, remove lr_readdir_msg from the request object
and remove its magic initialization.  Make callers provide
a msg structure that they initialize themselves.  This will
allow future code to inspect the contents of a directory
before it is sent over the wire.

I also added a bunch of comments.