backend/fs.c: re-fix directory reading
Now that I check the lstat() call for failure, the old style
9P2000 and 9P2000.u directory reads all fail, because the old
for providing qids was broken.  We are reading the directory
named in file->name and we get names within that directory.

As sef noted, we really want lstatat() here.  FreeBSD provides
it via fstatat() with flag AT_SYMLINK_NOFOLLOW.  MacOS does not,
so build full paths and use lstat() there.