utils.c: minor typos and type fussiness
When decoding a name, if it's NULL, print the prefix string
before the word <null>, don't pass the NULL again.

When decoding values with masks (some of which are actually
64 bit and some of which are 32 bit but we store in a 64-bit
variable), use a field width determined by the fcall bit width
but use PRIx64 since the variable itself is 64 bit.