rename l9p_openfile to l9p_fid
The struct name will be used a great deal more soon,
and it is currently both too long, and misleading.
A fid inside the server is not necessarily an *open*
file, as it can identify a not-yet-opened file or directory.
When we add extended attributes, it can identify an xattr,
which is not a file at all.