l9p_iov_io: fix fencepost error
At entry to l9p_iov_io, we correctly reject an attempt to
read from or write to the iov when the "next iov index" is
greater than or equal to the number of iov vectors.

Inside the loop, however, when we run out of the current
iov and advance to the next, we reject the I/O if we advance
the next-iov-index such that it is greater than the number
of iov vectors.  Thus, we go ahead and use iov[max] when the
correct range is iov[0] through iov[max - 1].

Probably we should return a short count here too, but hold
off on that until we vet callers.