move l9p_respond() call to requester
Change most of the backend interface from "void" to "int":
have all the operations return an error number, or 0 for
success, rather than calling l9p_respond() themselves.

Every request is ended with a response, but now the response
happens in the request's caller.

This is a large change, but mostly mechanical.  It is
intended to open the door to changing the backend interface
to pass operation-specific parameters in place of a single
"struct l9p_request *".  We will then be able to move any
request-specific details into request.c.