fix fs walk operation
I misunderstood some of the requirements for walk, and hence
broke it in commit 56637d296a8185b7c584f4bcadfd24c63d3e02dd.
We are not looking up each name in the same original single
directory, but rather treating each name component as a
normal file-or-directory as usual, until we attempt to walk
INTO it, when we treat it as a directory.

Note that because of the S_ISDIR test, we will refuse to walk
into a symlink in the server, forcing the client to handle
symbolic link resolution.  This is probably what we want for
directory handling, but perhaps we need to resolve them in
the server.

File handling of symbolic links (when opening them as files)
is clearly incorrect.  We either need to reject attempts to
open symlinks, or resolve such names in the server

If we must resolve symlinks in the server, we cannot simply
pass them on to the OS, since we are not doing a chroot() in
the server and hence the server's idea of the file system
root differs from the client's.