fix bugs in Tattach and remove lo_qid
Remove lo_qid from l9p_fid structure: the only code that
used it was in backend/fs.c fs_attach().

That code was wrong, along with the Tattach handler in
request.c that uses it.

Specifically, the fid in an attach request is by definition
a *new* fid.  If it is an in-use fid we should fail with
EINVAL, not look up an existing, in-use fid.  The previous
code would modify (and leak memory from) any existing fid
a client improperly supplied and the server successfully
attached, or clunk an existing fid improperly supplied and
failed-to-attach.

If we failed to attach because the user ID or name supplied
was incorrect, we leaked a "struct openfile".

If the (per-fs-softc) file system root was not, in fact, a valid
directory, we let the attach succeed but later operations would
fail (unless someone outside the server managed to create it as a
directory), because we constructed a fake directory qid and
assumed the path was valid.  Now we check that path refers to a
directory, and generate a real qid.

Log the reason for any failure -- we need to make sure this
works in both success and fail cases now.