utils: fix describe_bits multi-bit fields
The O_CREAT bit for Linux flags was missing, and I got:

    [DEBUG] l9p_dispatch_request: Tlcreate tag=1 fid=2
    name="file3" flags=0x8041[O_WRITE,O_LARGEFILE,?0x41]
    mode=8000<rw-r--r--> gid=0

Note the ?0x41 for "flags", which should read ?0x40.  We
need to clear the mask when any of the matching enumeration
values matches.  Simplest seems to be to clear immediately,
then make sure we'll skip additional same-mask values, so
that we don't accidentally match a zero-valued enumeration
for this same field.

(For single-bit fields the inner loop will not run at all.)