utils: retain permission bits
The debug message:

    ... Tlcreate ... mode=8000<rw-r--r-->

correctly shows that there's a mystery bit (0x8000) in the
permissions (aka "mode") field, but is equally mysterious
about (a) being in hexadecimal and (b) where the rw-r--r--
part came from.  This was due to a faulty format missing
the 0x prefix, and being too eager to clear "known" bits
from the original value.  Fix both.