fcall.h: add NOFID, NONUNAME; tweak l9p_stat types
The protocol describes particular values in Tattach
that mean "not a valid authentication fid, do not use"
and "not a valid n_uname, do not use".  Define macros
for them using the L9P_ prefix and the protocol's names.
Use the protocol-specified type (uint32_t) for both,
not one modified to fit our fancy (uid_t).

While here, fix l9p_stat n_uid, n_gid, and n_muid types
to be those in the protocol (uint32_t) as well.  This
is not as informational as uid_t and gid_t but is more
correct, especially if the bitsize of the local types
does not match that of the protocol.