backend/fs.c: restructure open, part 2
Move fs_open() and fs_lopen() common code into fs_iopen().
As a side effect this fixes 9P2000 and .u open to correctly
obey the read-only flag.

I put in a long comment as to what we are doing here, how we get a
bit more security from it, and how it fails to provide sufficient
security.  Essentially, clients can conspire to replace internal
path components with symlinks.

Suppose file system path /export/top is exported to clients A and
B.  Client A creates directory a/ and asks to create or open
a/passwd (whose server path is "/export/top/a/passwd").

Client B now races in, removes directory a/, and puts in a
symlink: "a" -> "/etc".

The server handling the request for Client A now uses the
previously vetted "/export/top/a/passwd" path.  Even with
O_NOFOLLOW, this happily follows the internal symlink "a"
and hence opens or creates "/etc/passwd" on the server.

The only cures for this are to use openat() heavily, so that
the server can be sure that "a/passwd" refers to the previously
verified "a" directory, or to add a kernel flag to forbid
internal (embedded) symlinks.