always set n_uname in attach, and fix minor type mixups
When using 9P2000 (not .u, not .L), set n_uname to
L9P_NONUNAME.  This way the backend code can always
look at n_uname to see if a numeric uname was provided,
regardless of protocol version.

In the backend, check for NONUNAME.  Use that to mean
"no numeric uname supplied", so that we don't depend on
it becoming (uid_t)-1.

Last, the group ID is type gid_t, not uid_t.