backend/fs.c: fix minor storage leak
If we run out of memory during attach handling,
we can leak the gid array allocated in l9p_getgrlist.
If we run out of memory just a bit earlier we can
pass a bogus value to malloc().  Neither of these
is likely to be a big deal, but fix them now, since I
noticed them in passing.