start marking fids valid and/or xattr
Simplify calls to the backend clunk and remove
function, which can take just a single fid now,
instead of a generic request.  Move the error
handling (disposing of temporary, never-became-valid
fids) out of the respond code, closer to where it
all makes sense.

Additional fid type-markings are yet to be implemented.
After that, we can actually start checking the validity
of incoming fids.