fix sbuf library usage
The debug output for Rwalk could print:

    [DEBUG] ... wqid=[<0x80[DIR],0,0x0000000000385e00>]s" aname=""

instead of:

    [DEBUG] ... wqid=[<0x80[DIR],0,0x0000000000385e00>]

on FreeBSD (but not on Mac where we used the private sbuf/sbuf.c
file).  This is because the FreeBSD routine to NUL-terminate the
sbuf is sbuf_finish(), not sbuf_done() (sbuf_done() is a predicate
returning true if sbuf_finish() was called).

Use the correct function, and rename the private version.