rewire lookup; allow empty component but not '/'
The protocol does not explicitly forbid empty name components,
it just never says what they mean.  Allow this test to succeed
either way, but retain it because it used to crash the server.

Add a test for literal '/', which the protocol *does* forbid.
We should test this with file / directory creation (through
all of Tcreate, Tlcreate, and Tmkdir) as well as for lookup,
but for the moment, just test lookup.