sequencer: improve debug
It's nice to be able to debug sequencer pack/unpack
operations when using doctest.testmod() from other
modules like protocol.py.  So, replace the simple
print-to-stdout with print-to-"debug" field, if debug
is not false-like.

Set each sequencer's debug to False, so that we print
nothing unless someone does, e.g.,

    fooseq.debug = sys.stderr

(which is what I did while debugging protocol.py changes).