fix mish-mash of dot-u and dot-L test issues
We were constructing 9P2000.u style stat data for
9P2000.L protocols if they invoked base 9P2000 requests
(such as read() on a directory, or Tstat operations).
This is officially wrong: the only dot-u style
message that dot-l shares is the Tauth n_uname.

Real Linux clients do not issue dot-u requests when
using dot-L, but this caused problems with the tests,
which do mix them together (on purpose).

In pytest/protocol.py, fix missing size field in
stat data.