pytest/client.py: clean up a failed test leftover
If we create dir/sub when we shouldn't be able to, try to
remove "sub" from "dir".  If the problem is like what happened
on NFS servers, this won't help, but it's worth a try.  (We
could also try using unlinkat with dir/sub, but only if the
client supports dot-L.)  If we can't even do that, just
stop here.