pytest/p9conn.py: fix fid items, uxopen, and write ops
When we allocate a fid for some operation like Twalk
that winds up totally failing, we should not clunk it,
but rather simply retire it.  (The dupfid case was missing
entirely, hence leaking a fid if it ever failed, not that
it should ever fail.)

When using uxopen to create a file, we got the fid manipulation
wrong, so all file create ops failed.

When a write failed (e.g., because caller stupidly opened the
file read-only), we tried to raise an error using a non-existent
variable.