pytest/: fix .L items and rework stat encoding
The stat operation does include .u extensions with .L ops
(Twstat etc), if they are used.  Likewise, Tattach and Tcreate,
if they are used in .L, take the .u extensions.  And, if
we somehow get or deliver Rerror with .u -- this should still
never happen -- we would need to include the error number
as with .u.

In other words, .L really is a strict superset of .u, except
that we "must not" use Rerror and "generally should not" (but
may, if we want to) use Twstat, read() on a directory fd, etc.
Simplify the special .uauth away.  Clean up the conditional
checking a little bit, while at it (this was a side effect of
expanding it temporarily during testing).

The on-wire stat does always have a prefix size field, in all
cases: Tread-a-directory, Rstat values, and Twstat requests.
The code for dealing with these was too painful, though, so
now we split it into the definition of a stat object (which
lacks the prefix size field) and a wirestat (which consists
of just a size, and data[size]) and do an explicit two-step
encode or decode as needed.