pytest/p9conn.py: fix several bugs in uxremove()
We were accidentally calling the remove op on the
root of a tree.  The Linux diod server then actually
did that remove.

We also could double-clunk or fail to clunk some fids.

To work with diod we must use the new style readdir,
not the old style read()-a-directory, so do that.