pytest/pyconn.py: fix numerous fid leaks
Drop the rootfid when shutting down.

Don't dup fids in the readdir code, we leaked them.  Sanitize
the code instead: the workhorses only open the fids to read the
directories, and do not clunk any; the callers then take care
of those operations.

This also helps sanitize the recursive remove as well.

In the uxcreate code, make sure we clunk the looked-up starting
directory when needed.  Move a bit of code to a function for
clarity.  While here, fix missing errno in OSError exception.