fix numerous remaining malloc()/free() issues
1. connection.c:

Free l9p_fid data structures in connection.c whenever we
discard the fid from the hash table.

2. hashtable.c:

Hashtable iterators were pretty thoroughly broken, and
never did anything (which was actually good).  For instance
initializing the iterator found the first item on slot 0,
if there was an item on slot 0 (if not, it would not iterate
over anything).  Then ht_next would skip that first item and
go on to the second item.

Besides that, ht_remove_at_iter removes the item we're
standing on.  This means the we must have a way to "preload"
the *next* item to return.  Add one, reworking the internal
iterator mechanisms.

The ht_destroy code failed to free any remaining hash items
(though the way we use the hash tables, there should never
be any).

3. lib9p.h, connection.c, request.c:

When we're done with a request, we need to strip it out of
the per-connection hash table, so that we won't iterate over
it at close, trying to free it.  Add a function to explicitly
release requests when we are done with them.

4.  transport/socket.c:

Release socket_softc data when the thread is done.