fix yet more storage issues and tag checking
We inserted requests into the hash table before decoding
them.  This meant we were using junk (well, zero) as the
tag every time.  Move the ht_add() call in connection.c
to *after* the packet-decode.

While at it, check for errors and free memory.  It's not
clear how to deal with failure to allocate a reply buffer
here, so we do that last, so that there is no remaining
failure mode to need a release-buffer from.

Move the responsibility for freeing a failed pufcall
to the caller, rather than having weird "if direction is..."
tests.

Move the responsibility for freeing the request itself
out of l9p_respond, into its callers, since one of them
(the hash table iterator) needs to be different.