threadpool.c: cleanup in error case; cosmetics
When we fail to create the threadpool, make sure we release the
mutex and condition variables.

While here, use a local variable in l9p_worker to make the
code more compact and readable.

This is a minor cleanup in preparation for handling Tflush better
(we'll be adding another condition variable and some more states).

Ticket: #19443