move transport functions and add "drop"
When we handle Tflush requests, we need the ability to drop
a request without replying to it.

We actually already need this for an error ("should not happen")
case, when a reply does not fit into the buffer reserved for it.
Right now this case, if it ever occurs, just leaks memory.

This commit will affect the bhyve transport code outside of
the library.  I collected the transport functions into their
own data structure (not technically required here) to guarantee
that compilation will fail if library-users do not update to
the new interface (and to simplify and document the setup).

Ticket: #19443