add Tflush support
This is a bit complicated.

When a Tflush arrives, we must find the request it is supposed
to flush.

If that request has been started, we should interrupt it if
necessary -- this is currently commented out -- and then wait
for it to finish.

If it has NOT been started, we should abort it quickly, and
wait for it to be handled.

If it does not exist, we simply acknowledge the flush request.

A request that is flushed (and not run) should normally be dropped
entirely, but the dropping mechanism is somewhat untested.  It's
valid to just reply to it with an error (e.g., EINTR) since this
is indistinguishable, on the client, from "never started", so for
now we do that.

A request that is flushed but completes successfully anyway (due
to racing between the flush and the request itself) must be
responded-to in full, as if the flush request has not yet been
received, and only then can the flush request be replied-to.

A flush that flushes the same request as an earlier flush could
be dropped, replying only to the *last* such flush, but that's
merely an efficiency hack.  For now we don't bother.

In all cases, the "flushee" (the request that is flushed) must
be fully dropped, or fully replied-to, before we answer the Tflush
request, and the Tflush itself always succeeds (if not dropped
entirely as part of the aforementioned efficiency hack).

To this end, we add a "flushed" state to every request (though
it is only used for non-Tflush requests), and (though it is used
only for flush requests) a "flushee done" flag.  When a flush
request (a "flusher") comes in, we handle it outside the thread pool,
finding the thread that handling the "flushee" if any.  We mark
the flushee and then wait for it to be responded-to.

This necessitated moving the tag hash lock around a bit and
letting the connection and request code hold it as needed.  We
also add a new condition variable, ltp_flush_cv, to the
threadpool, so that flushers can wait for flushees to mark]
their flushers.

Ticket: #19443