rework Tflush support yet again
The virtio version of 9p deadlocked with the tricky use of
the hash tag lock.  (It worked for the socket transport, but
that's not the case we care about.)

Instead, to solve the ordering of replies, we now have a
single thread dedicated to sending reply messages out of an
atomic queue.  We make sure to put Rflush replies to Tflush
requests into the queue only *after* the original request
has its reply (or drop) queued.

This is architecturally cleaner, as all the flush handling is
now localized to threadpool.c, and means we hold no locks
across the drop-or-reply calls.  We also now never hold up
the incoming request path: Tflush operations are queued on
their flushee, if there is one, and we then return immediately.

All the "drop" code is currently ifdef-ed out, except for the
one special case of a transport that provides buffers that are
too small to hold replies in the first place (which is a
pretty dreadful error: the protocol cannot work in this case).