backend/fs.c: fix readdir overrun
When packing readdir() dirent objects, it's possible for
l9p_pudirent() to fail (return negative value) after
it packs *some part* of the dirent.  In this case the
msg.lm_size value has become invalid.

Simply grab each valid size as we make progress, so
that we have the correct count when we exit the loop
for whatever reason.

(While here, fix up some comments.)