add debug logging for ACL checking
It's extra-noisy, so it is on its own ifdef.  (Need to
improve the debug facilities.)  Debug log showed that
these were working as intended; it was the OS's ACL inheritance
and settings that were not (FreeBSD ZFS NFSv4-style ACLs must
be set up through "zfs set aclmode=... aclinherit=... <dataset>").