use NFSv4 ACLs for access control
Convert everything in the backend to look for, and apply,
NFSv4 ACLs.  If the underlying file system does not support
NFSv4 ACLs, fall back on normal Unix style permissions.

We have a master disable in the authinfo as well, but at
the moment, no way to set it per connection.  So at the
moment, we use ACLs if the file system provides ACLs.