backend/fs.c: make ai_refcnt adjustments atomic
Use a mutex to make increment and decrement of ai_refcnt
atomic.  We have a demo-bug of two threads both incrementing
ai_refcnt and leaving it set to 8, leading to premature freeing:

[DEBUG]  open_fid: open_fid: authinfo 0x88560e380 now used by 8
[DEBUG]  open_fid: open_fid: authinfo 0x88560e380 now used by 8
[DEBUG]  l9p_respond: Rwalk tag=1 wqid=[]
[DEBUG]  l9p_respond: Rwalk tag=2 wqid=[<0x80[DIR],0,0x0000000000000015>]
   [snip]
[DEBUG]  l9p_dispatch_request: Tclunk tag=1 fid=8
[DEBUG]  fs_freefid: fs_freefid: dropped last ref to authinfo 0x88560e380
[DEBUG]  l9p_respond: Rclunk tag=1
   [snip]
[DEBUG]  l9p_dispatch_request: Twalk tag=1 fid=7 newfid=8 wname="Library"
[DEBUG]  l9p_check_aces: ACE: allow OWNER@: mask 0x20 ace_mask 0x1e1bf
[DEBUG]  l9p_check_aces: clear 0x1e1bf: now mask=0x0
[DEBUG]  l9p_check_aces: ACE: end of ACEs, mask now 0x0: ALLOW
[DEBUG]  open_fid: open_fid: authinfo 0x88560e380 now used by 1

Most of the authinfo is read-only, except for the ref count, which
we increment and decrement.  Now that we're adjusting the field in
free()d memory that gets re-used later, the code eventually
crashes.

While we're at it, we also assert that increments result in values
that exceed 1 (i.e., did not start from 0), except when creating
the initial authinfo.  This will only catch some cases, but the
above *is* one it would catch.

The mutex is not pretty, but ifdefs are worse, and this code path
is not critical for speed.

(While here, also move ff_flags for better packing of an fs_fid.)