move superuser semantics ahead of ACLs
In the generic ACL checker, check for superuser access before,
ACLs, if superuser is enabled at all.

In the pathname based backend, enable superuser semantics for
ACLs as well as for stat-only.

This is needed for Linux mmap() as it opens files as superuser
regardless of the uid/gid of the process opening the file.