request.c: log un-run flushed requests
If a request gets flushed without ever reaching the worker
threads, it never goes through l9p_dispatch_request and hence
never gets logged in the debug log code.  So, when we're
flushing such a request, let's show the original request
first.

(Maybe we should repeat the entire request for all flushed
requests.  But for now, just show those that are in this
state.)

(The theory behind this is that Linux is sending Tflush requests
for Tclunk requests, and we're aborting the Tclunk and hence
leaving a fid set valid that Linux thinks should now be invalid.)