prevent Tflush from flushing Tclunk
Linux client sends Tflush requests for Tclunk operations.
If we flush them without ever running them, as we were
doing, this leaves the corresponding fid open, but Linux
expects it to become closed.

The same logic applies to Tremove since it also clunks a
fid.  It might be desirable to split the remove and close
actions, allowing the remove to be bypassed, but that can be
done later.

For safety, this marks Tversion, Tattach, and Tflush as
must-run as well.

Ideally, we should still use signals to interrupt blocking
operations (including, perhaps, close, though it is not
clear how this should work) if they get a Tflush after being
dispatched, but for now this remains a future item.

Ticket: #21214