short circuit back end getlock
The Tgetlock operation is basically useless.  Qemu just
says "success", so we do too, now.  (Maybe this should
move up a level into request.c, but for now we'll just
do it here.)

The existing code was wrong: it never filled in the "fl" variable
before calling ioctl.  So the return error code was sort of
random, and if the call worked, the returned lock range was not
sensible.  But the operation is pointless:
 the Linux client looks only at the error code anyway;
any actual lock data only tells you what the lock state used to
be, which is no longer valid by the time you get it.